Update the loop index to long for RpcServer.Listener#cleanupConnections method
Summary: Update the loop index to long for RpcServer.Listener#cleanupConnections method

Test Plan: N/A

Reviewers: #hbase, zhangduo

Reviewed By: #hbase, zhangduo

Maniphest Tasks: T13590

Differential Revision: https://phabricator.d.xiaomi.net/D141530

+++ hbase-server/src/main/java/org/apache/hadoop/hbase/ipc/RpcServer.java
@@ -637,25 +637,25 @@ public class RpcServer implements RpcServerInterface {
         if (!force && (currentTime - lastCleanupRunTime) < cleanupInterval) {
           return;
         }
-        int start = 0;
-        int end = numConnections - 1;
+        long start = 0;
+        long end = numConnections - 1;
         if (!force) {
           start = rand.nextInt() % numConnections;
           end = rand.nextInt() % numConnections;
-          int temp;
+          long temp;
           if (end < start) {
             temp = start;
             start = end;
             end = temp;
           }
         }
-        int i = start;
+        long i = start;
         int numNuked = 0;
         while (i <= end) {
           Connection c;
           synchronized (connectionList) {
             try {
-              c = connectionList.get(i);
+              c = connectionList.get((int) i);
             } catch (Exception e) {return;}
           }
           if (c.timedOut(currentTime)) {
